+++ Sources/parser/OPFParser.swift
@@ -19,8 +19,11 @@ enum OPFParserError: Error {
 /// EpubParser support class, able to parse the OPF package document.
 /// OPF: Open Packaging Format.
 public class OPFParser {
+    let smilp: SMILParser!
 
-    internal init() {}
+    internal init() {
+        smilp = SMILParser()
+    }
 
     /// Parse the OPF file of the Epub container and return a `Publication`.
     /// It also complete the informations stored in the container.
@@ -145,7 +148,7 @@ public class OPFParser {
                 // Retrieve the duration of the smil file
                 if let duration = publication.metadata.otherMetadata.first(where: { $0.property == "#\(id)" })?.value {
 
-                    link.duration = Double(smilTimeToSeconds(duration))
+                    link.duration = Double(smilp.smilTimeToSeconds(duration))
                 }
                 //publication.links.append(link)
             }
@@ -189,7 +192,9 @@ public class OPFParser {
         }
     }
 
-    /// Parse the mediaOverlays documents.
+    /// Parse the mediaOverlays informations contained in the ressources then
+    /// parse the associted SMIL files to populate the MediaOverlays objects 
+    /// in each of the Spine's Links.
     ///
     /// - Parameters:
     ///   - container: The Epub Container.
@@ -203,7 +208,6 @@ public class OPFParser {
             log(level: .info, "No media-overlays found in the Publication.")
             return
         }
-        print("=++= \(publication.metadata.title)")
         for mediaOverlayLink in mediaOverlays {
             let node = MediaOverlayNode()
             let smilXml = try container.xmlDocument(forRessourceReferencedByLink: mediaOverlayLink)
@@ -212,8 +216,8 @@ public class OPFParser {
             node.role.append("section")
             node.text = body.attributes["epub:textref"]
             // get body parameters <par>
-            parseParameters(in: body, withParent: node)
-            parseSequences(in: body, withParent: node, publicationSpine: &publication.spine)
+            smilp.parseParameters(in: body, withParent: node)
+            smilp.parseSequences(in: body, withParent: node, publicationSpine: &publication.spine)
             // "../xhtml/mo-002.xhtml#mo-1" => "../xhtml/mo-002.xhtml"
 
             guard let baseHref = node.text?.components(separatedBy: "#")[0],
@@ -225,8 +229,6 @@ public class OPFParser {
             }) else {
                 continue
             }
-
-            print("=++= DANS UPPER on ajoute du media overlay a \(String(describing: link.href))")
             link.mediaOverlays.append(node)
             link.properties.append(EpubConstant.mediaOverlayURL + link.href!)
         }
@@ -268,213 +270,6 @@ public class OPFParser {
         return link
     }
 }
-
-// MARK: - Media Overlays.
-extension OPFParser {
-
-    /// [RECURSIVE]
-    /// Parse the <seq> elements at the current XML level. It will recursively
-    /// parse they childrens <par> and <seq>
-    ///
-    /// - Parameters:
-    ///   - element: The XML element which should contain <seq>.
-    ///   - parent: The parent MediaOverlayNode of the "to be creatred" nodes.
-    fileprivate func parseSequences(in element: AEXMLElement,
-                                    withParent parent: MediaOverlayNode,
-                                    publicationSpine spine: inout [Link])
-    {
-        guard let sequenceElements = element["seq"].all,
-            !sequenceElements.isEmpty else
-        {
-            return
-        }
-        // TODO: 2 lines differ from the version used in the parseMediaOverlay
-        //       for loop. Refactor
-        for sequence in sequenceElements {
-            let newNode = MediaOverlayNode()
-
-            newNode.role.append("section")
-            newNode.text = sequence.attributes["epub:textref"]
-            parseParameters(in: sequence, withParent: newNode)
-            parseSequences(in: sequence, withParent: newNode, publicationSpine: &spine)
-
-            let baseHrefParent = parent.text?.components(separatedBy: "#")[0]
-
-            guard let baseHref = newNode.text?.components(separatedBy: "#")[0],
-                baseHref != baseHrefParent else
-            {
-                parent.children.append(newNode)
-                continue
-            }
-            guard let link = spine.first(where: {
-                guard let linkRef = $0.href else {
-                    return false
-                }
-                return linkRef.contains(baseHref) || baseHref.contains(linkRef)
-            }) else {
-                continue
-            }
-            print("=++= LOWER on ajoute du media overlay a \(String(describing: link.href))")
-            link.mediaOverlays.append(newNode)
-            link.properties.append(EpubConstant.mediaOverlayURL + link.href!)
-        }
-    }
-
-    /// Parse the <par> elements at the current XML element level.
-    ///
-    /// - Parameters:
-    ///   - element: The XML element which should contain <par>.
-    ///   - parent: The parent MediaOverlayNode of the "to be creatred" nodes.
-    fileprivate func parseParameters(in element: AEXMLElement,
-                                     withParent parent: MediaOverlayNode)
-    {
-        guard let parameterElements = element["par"].all,
-            !parameterElements.isEmpty else
-        {
-            return
-        }
-        // For each <par> in the current scope.
-        for parameterElement in parameterElements {
-            let newNode = MediaOverlayNode()
-
-            guard let audioElement = parameterElement["audio"].first else {
-                return
-            }
-            let audioFilePath = parse(audioElement: audioElement)
-
-            newNode.audio = audioFilePath
-            newNode.text = parameterElement["text"].attributes["src"]
-            parent.children.append(newNode)
-        }
-    }
-
-    /// Converts a smile time string into seconds String.
-    ///
-    /// - Parameter time: The smile time String.
-    /// - Returns: The converted value in Seconds as String.
-    fileprivate func smilTimeToSeconds(_ time: String) -> String {
-        let timeFormat: SmilTimeFormat
-
-        if time.contains("h") {
-            timeFormat = .hour
-        } else if time.contains("s") {
-            timeFormat = .second
-        } else if time.contains("ms") {
-            timeFormat = .milisecond
-        } else {
-            let timeArity = time.components(separatedBy: ":").count
-
-            guard let format = SmilTimeFormat(rawValue: timeArity) else {
-                return ""
-            }
-            timeFormat = format
-        }
-        return timeFormat.convertToseconds(smilTime: time)
-    }
-
-    /// Parse the <audio> XML element, children of <par> elements.
-    ///
-    /// - Parameter audioElement: The audio XML element.
-    /// - Returns: The formated string representing the data.
-    fileprivate func parse(audioElement: AEXMLElement) -> String? {
-        guard var audio = audioElement.attributes["src"],
-            let clipBegin = audioElement.attributes["clipBegin"],
-            let clipEnd = audioElement.attributes["clipEnd"] else
-        {
-            return nil
-        }
-        /// Clean relative path elements "../"
-        let components = audio.components(separatedBy: "/")
-        if components[0] == ".." {
-            audio.removeSubrange(audio.startIndex..<audio.index(audio.startIndex, offsetBy: 3))
-        }
-        audio += "#t="
-        audio += smilTimeToSeconds(clipBegin)
-        audio += ","
-        audio += smilTimeToSeconds(clipEnd)
-        return audio
-    }
-}
-
-/// Describe the differents time string format of the smile tags.
-///
-/// - splitMonadic: Handle `SS` format.
-/// - splitDyadic//MM/SS: Handles `MM/SS` format.
-/// - splitTriadic//HH:MM:SS: Handles `HH:MM:SS` format.
-/// - milisecond: Handles `MM"ms"` format.
-/// - second: Handles `SS"s" || SS.MM"s"` format
-/// - hour: Handles `HH"h" || HH.MM"h"` format.
-fileprivate enum SmilTimeFormat: Int {
-    case splitMonadic = 1
-    case splitDyadic
-    case splitTriadic
-    case milisecond
-    case second
-    case hour
-}
-
-fileprivate extension SmilTimeFormat {
-
-    /// Return the seconds double value from a possible SS.MS format.
-    ///
-    /// - Parameter seconds: The seconds String.
-    /// - Returns: The translated Double value.
-    fileprivate func parseSeconds(_ time: String) -> Double {
-        let secMilsec = time.components(separatedBy: ".")
-        var seconds = 0.0
-
-        if secMilsec.count == 2 {
-            seconds = Double(secMilsec[0]) ?? 0.0
-            seconds += (Double(secMilsec[1]) ?? 0.0) / 1000.0
-        } else {
-            seconds = Double(time) ?? 0.0
-        }
-        return seconds
-    }
-
-    /// Will confort the `smileTime` to the equivalent in seconds given it's
-    /// type.
-    ///
-    /// - Parameter time: The `smilTime` `String`.
-    /// - Returns: The converted value in seconds.
-    func convertToseconds(smilTime time: String) -> String {
-        var seconds = 0.0
-
-        switch self {
-        case .milisecond:
-            let ms = Double(time.replacingOccurrences(of: "ms", with: ""))
-            seconds = (ms ?? 0) / 1000.0
-        case .second:
-            seconds = Double(time.replacingOccurrences(of: "s", with: "")) ?? 0
-        case .hour:
-            let hourMin = time.replacingOccurrences(of: "h", with: "").components(separatedBy: ".")
-            let hoursToSeconds = (Double(hourMin[0]) ?? 0) * 3600.0
-            let minutesToSeconds = (Double(hourMin[1]) ?? 0) * 0.6 * 60.0
-
-            seconds = hoursToSeconds + minutesToSeconds
-        case .splitMonadic:
-            return time
-        case .splitDyadic:
-            let minSec = time.components(separatedBy: ":")
-
-            // Min
-            seconds += (Double(minSec[0]) ?? 0.0) * 60
-            // Sec
-            seconds += parseSeconds(minSec[1])
-        case .splitTriadic:
-            let hourMinSec = time.components(separatedBy: ":")
-
-            // Hour
-            seconds += (Double(hourMinSec[0]) ?? 0.0) * 3600.0
-            // Min
-            seconds += (Double(hourMinSec[1]) ?? 0.0) * 60
-            // Sec
-            seconds += parseSeconds(hourMinSec[2])
-        }
-        return String(seconds)
-    }
-}
-
